fix repack/promptreco task archive problems
There is currently a race condition between
Repack, PromptReco and the TaskArchiver. If
repacking is done before PromptReco is
configured, it looks like everything is
finished and the TaskArchiver will try to
cleanup repacking, including the fileset
that is the input for PromptReco.

To prevent the TaskArvchiver from seeing the
repack workflow, only mark that workflow
as injected after all primary datasets
for the stream that is repacked have had
their PromptReco configured and released.

Also included a CASCADE DELETE in the Tier0
schema on the assoc table that holds the
input fileset for PromptReco.