add PCL support to Express, fixes #3567
When a run is configured, any express streams with the PromptCalibProducer alca producer will
be flagged for PCL. This means the system expects PCL conditons for that run/stream pair.
Inside the Express Spec we configure for ALCAPROMPT output from the PromptCalibProducer and
then add an AlcaHarvest task for it. The AlcaHarvest will write out sqlite files, for which
we add a dummy Condition task. A Condition task has a Condition job splitter, which does
not create any jobs, but just forwards sqlite file information to special t0ast db tables.

Second step (and not part of this ticket) is using the information in these t0ast tables
to upload the conditions to the dropbox and enforce doing this in run order.

Also not part of this ticket yet is a timeout trigger for the PCL.