fix bug in lumi insertion
Had a crash where a lumi was passed twice in the same set of bind variables
to the lumi insert code, this caused a race condition and crash. Underlying
cause was that we find new closed lumis from the SM DB per stream but then
insert them stream independent. We missed filtering out duplicates.