fix bug in the prompt calibration loop
Two types of bugs in the PCL. We passed a stream id instead of
a stream name to a call. Plus a logic bug that prevented this
call from being made in almost all cases (this is how it slipped
through testing). Put in a temporary fix that will restore
upload functionality, but isn't final yes.