more adjustments for the prompt calibration loop
Change PCL logic. Instead of having three states only have two now,
finished and not finished. The AlcaHarvest triggers on a closed fileset
and (once) on timeout. The ConditionUpload will upload everything,
but only set the PCL to finished if it's done. The later is detected
by checking the status of the subscription. The input fileset needs
to be closed and there can be no available or acquired files.