fix problem with PCL completion check, fixes #4013
If AlcaHarvest timeout triggered and caused conditions to be created
for upload before all the data was processed, we need to make sure
that the system would only upload these conditions and not set
the PCL to finished for this run. For this reason there is the
completion check in the ConditionUpload.IsPromptCalibrationFinished
DAO. The query in that DAO has a bug, it would always return true.
The patch fixes the query.