fix some problems with handling of empty lumis in repackmerge
The repackmerge job splitting algorithm enforces lumi order and
only allows holes in the lumi order when there is no streamer
files for that lumi. This is not the only coniditon for which
we don't get RAW data though, it can also be that for that lumi
no events were selected for the dataset in question. Add a
check for that as well. Add a unittest to test this case.

Also simplify the lumi hole detection algorithm a bit and
apply the same simplification to the repack job splitting.