fix problem in lumi closing query
I added a consistency to the discovery of closed lumis
in the SM database and tried to be clever by integrating
it into the discovery query. Turns out I was too clever
as the new query has a problem and doesn't find all
closed lumis. Revert to the previous discovery version
query and add the check as an extra query in the same DAO.
Maybe not as efficient, but it's safer and it's also the
same way this is implemented in the old Tier0.