change the way PhEDex subscriptions are made
With the disk/tape separation and a new data placement model, the Tier0 will have to make different type of subscriptions to CERN tape and T1 tape and disk. Implement these changes and also simplify the configuration to only specify the CERN site (archival_node) and the T1 tape and disk endpoints (tape_node and disk_node).