deal with case of stream injected but not in HLT menu
If we feed data for a stream that is not part of the HLT menu of if there is a stream that has not datasets defined in the HLT menu, RunConfig will already crash. The error message is misleading though, add a separate check for this and fail with a proper message.