put a rate limit on express processing
Express goal is a low latency processing. If the rate is too high this goal is not achievable and worse, it can completely clog the system also for lower rate data.

Add a rate limit, above which a lumi will not be processed at all. The limit is configurable, default is 23000 events per lumi (with 23s lumis that is 1000Hz).