mark run/stream as done quicker
Currently run/stream is marked done when a run/stream workflow is archived and cleaned up. Due to WMStats retention policy this takes a week, which is a bit long. Change this to mark a run/stream as done if all subscriptions for this run/stream are marked as finished.